+++ CHANGELOG.md
@@ -10,7 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: At xxs, use thinner gutters.
 - Fix: Override bootstrap on blockquotes in `k-Testimony__text`
   and `k-PressQuotes__quote__text`.
-- Feature: Add new `initialTouched` prop to `LoanSimulator`.
+- Feature: Add new `initialTouched` and `requiredDurationError` props
+  to `LoanSimulator`.
 
 ## [5.2.0] - 2017-01-12